+++ repoze/workflow/security.py
@@ -1,0 +1,5 @@
+def has_permission(permission, context, request):
+    """ Default permission checker """
+    return request.has_permission(
+        permission,
+        context=context)